Add NetBSD support for Additions in common
Beside for the necessnary changes in several Makefile.kmk, implements a vboxguest driver module for NetBSD is hard and confused me a lot. After Scan the FreeBSD and Solaris implemations and according to the /usr/src/sys/dev/ files I implements a sample driver. No test have been done, so it's just a daemon.

Signed-off-by: Haomai Wang <haomaiwang@gmail.com>